Tickets - change decimal to int

+++ csharp-refactoring/RefactoringCouplers/RefactoringCouplers/FeatureEnvy/Correct/Ticket.cs
@@ -15,7 +15,7 @@ public class Ticket
         return Available;
     }
 
-    private bool CanAfford(decimal points)
+    private bool CanAfford(int points)
     {
         return points >= Price;
     }
